Holy fuck I've just done a closure!! there awesome!
what actually has happened here is i made a thing to find where the
client is, check the web service for bus stops and plot them on a map.
Added event listener to marker objects to pull the arrival details for
that stop. Hard coded at 88 for now. Thats the most important one